pyproject.toml
[project]
name = "django-table-ui"
version = "0.1.0"
description = "Experiments with rendering manageable tables from django models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
]

[dependency-groups]
dev = [
    "django-click>=2.4.0",
    "django-ninja>=1.3.0",
    "factory-boy>=3.3.1",
    "ruff>=0.9.2",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D212", "D203", "D102", "D105"]

[tool.ruff]
line-length = 120
